Fixed Capture Issues
The id and header_id have now all been changed to BIGINTs.
The srcaddr, dstaddr, srcport, and dstport fields have all been updated
to be UNSIGNED. This is causing an issue with the dstaddr field now
going out of range. Will look into it.
PacketV5 now correctly inserts in the PACKET_V5 table. It was missing a
'+ ", " +' between two fields.